Move NeuralNetClassifier and -Regressor to own modules.
* adjust imports to reflect that
* deprecate old imports
