+++ skorch/tests/test_callbacks.py
@@ -246,7 +246,7 @@ class TestCheckpoint:
     @pytest.fixture
     def net_cls(self):
         """very simple network that trains for 10 epochs"""
-        from skorch.net import NeuralNetRegressor
+        from skorch import NeuralNetRegressor
         import torch
 
         class Module(torch.nn.Module):
@@ -328,7 +328,7 @@ class TestProgressBar:
     @pytest.fixture
     def net_cls(self):
         """very simple network that trains for 2 epochs"""
-        from skorch.net import NeuralNetRegressor
+        from skorch import NeuralNetRegressor
         import torch
 
         class Module(torch.nn.Module):
